Initial Commit, Reading Config Struct and Documentation
This is a far-from-functional revision.  The only thing I've
implemented so far is the reflection code to read the field names,
types and tags from the configuration struct passed into the
ReadConfig function.  I've also written in some of the autodoc
comments for how conflag should work when it's finished.