Added concreteConfig.findConfigFile
This adds logic to load up the appropriate config file before parsing
options.  This entails scanning through the command-line flags to
check for possible config file names, possibly closing an
already-opened default file if one is found, and overriding the
default config file options.  I've included a unit test for it as
well.