Removed args Param From findConfigFile()
I realized that the findConfigFile() method on the concreteConfig
struct was expecting an args parameter passed into it.  Since the args
for the configuration are stored on the struct, that didn't really
make any sense.  I've adjusted the function to just use the config's
specified args and the unit test to set them appropriately.