Added Usage Text Generation
This adds a Usage method to the Configuration object, which returns
usage text for the program formatted for a given terminal width.  The
formatter will include program name and description, field flags and
file keys, and descriptions of each field (side-by-side with the keys
and flags if possible, otherwise beneath them).