Change Config and Field to structs
I initially made Config and Field interfaces with hidden structs that
implemented them.  The idea was to make it more testable, but I've
since come to see it as just overengineering.  This replaces them with
simpler struct definitions, and should have no impact on the public
API except that in lieu of Config and Field, client code will now be
interacting with *Config and *Field.